Refactoring done, rossocekt relay is tested, drone is tested for full funcationality

+++ src/game_master/scripts/rossocket_relay.py
@@ -103,7 +103,7 @@ class RosProxySocketListener:
                             print("Invalid command: {0}".format(self.buffer))
                         target = Twist()
                         target.linear.x     = float(cmd[1])
-                        target.linear.y     = int(cmd[2])
+                        target.linear.y     = float(cmd[2])
                         target.angular.z    = float(cmd[3])
                         self.target_pub.publish(target)
                     else:
